+++ src/systems/a320_systems_wasm/src/lib.rs
@@ -2,6 +2,7 @@
 mod ailerons;
 mod autobrakes;
 mod brakes;
+mod elevators;
 mod flaps;
 mod nose_wheel_steering;
 
@@ -9,6 +10,7 @@ use a320_systems::A320;
 use ailerons::ailerons;
 use autobrakes::autobrakes;
 use brakes::brakes;
+use elevators::elevators;
 use flaps::flaps;
 use nose_wheel_steering::nose_wheel_steering;
 use std::error::Error;
@@ -156,6 +158,7 @@ async fn systems(mut gauge: msfs::Gauge) -> Result<(), Box<dyn Error>> {
             .with_aspect(nose_wheel_steering)?
             .with_aspect(flaps)?
             .with_aspect(ailerons)?
+            .with_aspect(elevators)?
             .build(A320::new)?;
 
     while let Some(event) = gauge.next_event().await {